+++ CHANGELOG.md
@@ -17,5 +17,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Dialog when installing file from CurseForge not showing progress
 - Checking for mods on CurseForge running out of memory causing issue
 - Reinstalling or Updating CurseForge mods hanging when calling api
+- Setup dialog not having correct spacing
 
 ### Misc